fix(GiniHealthSDK): Split XCTests in multiple files to reduce file length
IPC-572
