+++ HealthSDK/GiniHealthSDK/Tests/GiniHealthSDKTests/GiniHealthTests.swift
@@ -23,6 +23,7 @@ final class GiniHealthTests: XCTestCase {
     }
 
     override func tearDown() {
+        giniHealthAPI = nil
         giniHealth = nil
         super.tearDown()
     }
@@ -38,7 +39,7 @@ final class GiniHealthTests: XCTestCase {
         XCTAssertEqual(GiniHealthConfiguration.shared, configuration)
     }
     
-    func testFetchBankingApps_Success() {
+    func testFetchBankingAppsSuccess() {
         // Given
         let expectedProviders: [GiniHealthSDK.PaymentProvider]? = loadProviders(fileName: "providers")
 
@@ -61,261 +62,8 @@ final class GiniHealthTests: XCTestCase {
         XCTAssertEqual(receivedProviders?.count, expectedProviders?.count)
         XCTAssertEqual(receivedProviders, expectedProviders)
     }
-
-    func testDocumentIsPayable() {
-        // When
-        let fileName = "extractionResultWithIBAN"
-        let extractions: GiniHealthSDK.ExtractionsContainer? = GiniHealthSDKTests.load(fromFile: fileName)
-        guard let extractions else {
-            XCTFail("Error loading file: `\(fileName).json`")
-            return
-        }
-        let extractionsResult = ExtractionResult(extractionsContainer: extractions)
-        let isPayable = extractionsResult.extractions.first(where: { $0.name == ExtractionType.paymentState.rawValue })?.value == GiniHealthSDK.PaymentState.payable.rawValue
-        // Then
-        XCTAssertEqual(isPayable, true)
-    }
     
-    func testDocumentIsNotPayable_Success() {
-        // When
-        let fileName = "extractionResultWithoutIBAN"
-        let extractions: GiniHealthSDK.ExtractionsContainer? = GiniHealthSDKTests.load(fromFile: fileName)
-        guard let extractions else {
-            XCTFail("Error loading file: `\(fileName).json`")
-            return
-        }
-        let extractionsResult = ExtractionResult(extractionsContainer: extractions)
-        let isPayable = extractionsResult.extractions.first(where: { $0.name == ExtractionType.paymentState.rawValue })?.value == GiniHealthSDK.PaymentState.payable.rawValue
-        // Then
-        XCTAssertEqual(isPayable, false)
-    }
-
-    func testCheckIfDocumentIsPayable_Success() {
-        // Given
-        let fileName = "extractionResultWithIBAN"
-        let expectedExtractions: GiniHealthSDK.ExtractionsContainer? = GiniHealthSDKTests.load(fromFile: fileName)
-        guard let expectedExtractions else {
-            XCTFail("Error loading file: `\(fileName).json`")
-            return
-        }
-        let expectedExtractionsResult = GiniHealthSDK.ExtractionResult(extractionsContainer: expectedExtractions)
-        let expectedIsPayable = expectedExtractionsResult.extractions.first(where: { $0.name == "iban" })?.value.isNotEmpty
-
-        // When
-        let expectation = self.expectation(description: "Checking if document is payable")
-        var isDocumentPayable: Bool?
-        giniHealth.checkIfDocumentIsPayable(docId: MockSessionManager.payableDocumentID) { result in
-            switch result {
-            case .success(let isPayable):
-                isDocumentPayable = isPayable
-            case .failure(_):
-                isDocumentPayable = nil
-            }
-            expectation.fulfill()
-        }
-        waitForExpectations(timeout: 1, handler: nil)
-
-        // Then
-        XCTAssertEqual(expectedIsPayable, isDocumentPayable)
-    }
-
-    func testCheckIfDocumentIsNotPayable_Success() {
-        // Given
-        let fileName = "extractionResultWithIBAN"
-        let expectedExtractions: GiniHealthSDK.ExtractionsContainer? = GiniHealthSDKTests.load(fromFile: fileName)
-        guard let expectedExtractions else {
-            XCTFail("Error loading file: `\(fileName).json`")
-            return
-        }
-        let expectedExtractionsResult = ExtractionResult(extractionsContainer: expectedExtractions)
-        let expectedIsPayable = expectedExtractionsResult.extractions.first(where: { $0.name == "iban" })?.value.isEmpty
-
-        // When
-        let expectation = self.expectation(description: "Checking if document is not payable")
-        var isDocumentPayable: Bool?
-        giniHealth.checkIfDocumentIsPayable(docId: MockSessionManager.notPayableDocumentID) { result in
-            switch result {
-            case .success(let isPayable):
-                isDocumentPayable = isPayable
-            case .failure(_):
-                isDocumentPayable = nil
-            }
-            expectation.fulfill()
-        }
-        waitForExpectations(timeout: 1, handler: nil)
-
-        // Then
-        XCTAssertEqual(expectedIsPayable, isDocumentPayable)
-    }
-
-    func testCheckIfDocumentIsPayable_Failure() {
-        // When
-        let expectation = self.expectation(description: "Checking if request fails")
-        var isDocumentPayable: Bool?
-        giniHealth.checkIfDocumentIsPayable(docId: MockSessionManager.failurePayableDocumentID) { result in
-            switch result {
-            case .success(let isPayable):
-                isDocumentPayable = isPayable
-            case .failure(_):
-                isDocumentPayable = nil
-            }
-            expectation.fulfill()
-        }
-        waitForExpectations(timeout: 1, handler: nil)
-
-        // Then
-        XCTAssertNil(isDocumentPayable)
-    }
-
-    func testCheckIfDocumentContainMultipleInvoices_Success() {
-        // When
-        let expectation = self.expectation(description: "Checking if document contains multiple invoices")
-        var hasMultipleInvoices: Bool?
-        giniHealth.checkIfDocumentContainsMultipleInvoices(docId: MockSessionManager.notPayableDocumentID) { result in
-            switch result {
-            case .success(let containsMultipleDocs):
-                hasMultipleInvoices = containsMultipleDocs
-            case .failure(_):
-                hasMultipleInvoices = nil
-            }
-            expectation.fulfill()
-        }
-        waitForExpectations(timeout: 5, handler: nil)
-
-        // Then
-        XCTAssertEqual(true, hasMultipleInvoices)
-    }
-
-    func testCheckIfDocumentDontContainMultipleInvoices_Success() {
-        // When
-        let expectation = self.expectation(description: "Checking if document don't contain multiple invoices")
-        var hasMultipleInvoices: Bool?
-        giniHealth.checkIfDocumentContainsMultipleInvoices(docId: MockSessionManager.payableDocumentID) { result in
-            switch result {
-            case .success(let containsMultipleDocs):
-                    hasMultipleInvoices = containsMultipleDocs
-            case .failure(_):
-                    hasMultipleInvoices = nil
-            }
-            expectation.fulfill()
-        }
-        waitForExpectations(timeout: 1, handler: nil)
-
-        // Then
-        XCTAssertEqual(false, hasMultipleInvoices)
-    }
-
-    func testCheckIfDocumentContainMultipleInvoices_Failure() {
-        // When
-        let expectation = self.expectation(description: "Checking if request fails")
-        var hasMultipleInvoices: Bool?
-        giniHealth.checkIfDocumentContainsMultipleInvoices(docId: MockSessionManager.failurePayableDocumentID) { result in
-            switch result {
-            case .success(let containsMultipleDocs):
-                hasMultipleInvoices = containsMultipleDocs
-            case .failure(_):
-                hasMultipleInvoices = nil
-            }
-            expectation.fulfill()
-        }
-        waitForExpectations(timeout: 1, handler: nil)
-
-        // Then
-        XCTAssertNil(hasMultipleInvoices)
-    }
-
-    func testPollDocument_Success() {
-        // Given
-        let healthDocument: GiniHealthAPILibrary.Document = GiniHealthSDKTests.load(fromFile: "document1")!
-        let expectedDocument: GiniHealthSDK.Document? = GiniHealthSDK.Document(healthDocument: healthDocument)
-
-        // When
-        let expectation = self.expectation(description: "Polling document")
-        var receivedDocument: GiniHealthSDK.Document?
-        giniHealth.pollDocument(docId: MockSessionManager.payableDocumentID) { result in
-            switch result {
-            case .success(let document):
-                receivedDocument = document
-            case .failure(_):
-                receivedDocument = nil
-            }
-            expectation.fulfill()
-        }
-        waitForExpectations(timeout: 1, handler: nil)
-
-        // Then
-        XCTAssertNotNil(receivedDocument)
-        XCTAssertEqual(receivedDocument, expectedDocument)
-    }
-    
-    func testPollDocument_Failure() {
-        // When
-        let expectation = self.expectation(description: "Polling failure document")
-        var receivedDocument: GiniHealthSDK.Document?
-        giniHealth.pollDocument(docId: MockSessionManager.missingDocumentID) { result in
-            switch result {
-            case .success(let document):
-                receivedDocument = document
-            case .failure(_):
-                receivedDocument = nil
-            }
-            expectation.fulfill()
-        }
-        waitForExpectations(timeout: 1, handler: nil)
-
-        // Then
-        XCTAssertNil(receivedDocument)
-    }
-    
-    func testGetExtractions_Success() {
-        // Given
-        let fileName = "extractionsWithPayment"
-        let expectedExtractionContainer: GiniHealthSDK.ExtractionsContainer? = GiniHealthSDKTests.load(fromFile: fileName)
-        guard let expectedExtractionContainer else {
-            XCTFail("Error loading file: `\(fileName).json`")
-            return
-        }
-        let expectedExtractions: [GiniHealthSDK.Extraction] = ExtractionResult(extractionsContainer: expectedExtractionContainer).payment?.first ?? []
-
-        // When
-        let expectation = self.expectation(description: "Getting extractions")
-        var receivedExtractions: [GiniHealthSDK.Extraction]?
-        giniHealth.getExtractions(docId: MockSessionManager.extractionsWithPaymentDocumentID) { result in
-            switch result {
-            case .success(let extractions):
-                receivedExtractions = extractions
-            case .failure(_):
-                receivedExtractions = nil
-            }
-            expectation.fulfill()
-        }
-        waitForExpectations(timeout: 1, handler: nil)
-
-        // Then
-        XCTAssertNotNil(receivedExtractions)
-        XCTAssertEqual(receivedExtractions?.count, expectedExtractions.count)
-    }
-    
-    func testGetExtractions_Failure() {
-        // When
-        let expectation = self.expectation(description: "Extraction failure")
-        var receivedExtractions: [GiniHealthSDK.Extraction]?
-        giniHealth.getExtractions(docId: MockSessionManager.failurePayableDocumentID) { result in
-            switch result {
-            case .success(let extractions):
-                receivedExtractions = extractions
-            case .failure(_):
-                receivedExtractions = nil
-            }
-            expectation.fulfill()
-        }
-        waitForExpectations(timeout: 1, handler: nil)
-
-        // Then
-        XCTAssertNil(receivedExtractions)
-    }
-    
-    func testCreatePaymentRequest_Success() {
+    func testCreatePaymentRequestSuccess() {
         // Given
         let expectedPaymentRequestID = MockSessionManager.paymentRequestId
 
@@ -339,7 +87,7 @@ final class GiniHealthTests: XCTestCase {
         XCTAssertEqual(receivedRequestId, expectedPaymentRequestID)
     }
     
-    func testOpenLink_Success() {
+    func testOpenLinkSuccess() {
         let mockUIApplication = MockUIApplication(canOpen: true)
         let urlOpener = URLOpener(mockUIApplication)
         let waitForWebsiteOpen = expectation(description: "Link was opened")
@@ -352,7 +100,7 @@ final class GiniHealthTests: XCTestCase {
         waitForExpectations(timeout: 0.1, handler: nil)
     }
     
-    func testOpenLink_Failure() {
+    func testOpenLinkFailure() {
         let mockUIApplication = MockUIApplication(canOpen: false)
         let urlOpener = URLOpener(mockUIApplication)
         let waitForWebsiteOpen = expectation(description: "Link was not opened")
@@ -363,165 +111,6 @@ final class GiniHealthTests: XCTestCase {
         })
 
         waitForExpectations(timeout: 0.1, handler: nil)
-    }
-    
-    func testSetDocumentForReview_Success() {
-        // Given
-        let fileName = "extractionsWithPayment"
-        let expectedExtractionContainer: GiniHealthSDK.ExtractionsContainer? = GiniHealthSDKTests.load(fromFile: fileName)
-        guard let expectedExtractionContainer else {
-            XCTFail("Error loading file: `\(fileName).json`")
-            return
-        }
-        let expectedExtractions: [GiniHealthSDK.Extraction] = ExtractionResult(extractionsContainer: expectedExtractionContainer).payment?.first ?? []
-
-        // When
-        let expectation = self.expectation(description: "Setting document for review")
-        var receivedExtractions: [GiniHealthSDK.Extraction]?
-        giniHealth.setDocumentForReview(documentId: MockSessionManager.extractionsWithPaymentDocumentID) { result in
-            switch result {
-            case .success(let extractions):
-                receivedExtractions = extractions
-            case .failure(_):
-                receivedExtractions = nil
-            }
-            expectation.fulfill()
-        }
-        waitForExpectations(timeout: 1, handler: nil)
-
-        // Then
-        XCTAssertNotNil(receivedExtractions)
-        XCTAssertEqual(receivedExtractions?.count, expectedExtractions.count)
-    }
-    
-    func testFetchDataForReview_Success() {
-        // Given
-        let fileName = "extractionsWithPayment"
-        let expectedExtractionContainer: GiniHealthSDK.ExtractionsContainer? = GiniHealthSDKTests.load(fromFile: fileName)
-        guard let expectedExtractionContainer else {
-            XCTFail("Error loading file: `\(fileName).json`")
-            return
-        }
-        let expectedExtractions: [GiniHealthSDK.Extraction] = ExtractionResult(extractionsContainer: expectedExtractionContainer).payment?.first ?? []
-        let documentFileName = "document4"
-
-        let healthDocument: GiniHealthAPILibrary.Document = GiniHealthSDKTests.load(fromFile: documentFileName)!
-        let expectedDocument: GiniHealthSDK.Document? = GiniHealthSDK.Document(healthDocument: healthDocument)
-
-        guard let expectedDocument else {
-            XCTFail("Error loading file: `\(documentFileName).json`")
-            return
-        }
-        let expectedDatForReview = DataForReview(document: expectedDocument, extractions: expectedExtractions)
-
-        // When
-        let expectation = self.expectation(description: "Fetching data for review")
-        var receivedDataForReview: DataForReview?
-        giniHealth.fetchDataForReview(documentId: MockSessionManager.extractionsWithPaymentDocumentID) { result in
-            switch result {
-            case .success(let dataForReview):
-                receivedDataForReview = dataForReview
-            case .failure(_):
-                receivedDataForReview = nil
-            }
-            expectation.fulfill()
-        }
-        waitForExpectations(timeout: 1, handler: nil)
-
-        // Then
-        XCTAssertNotNil(receivedDataForReview)
-        XCTAssertEqual(receivedDataForReview?.document, expectedDatForReview.document)
-        XCTAssertEqual(receivedDataForReview?.extractions.count, expectedDatForReview.extractions.count)
-    }
-    
-    func testFetchDataForReview_Failure() {
-        // When
-        let expectation = self.expectation(description: "Failure fetching data for review")
-        var receivedError: GiniHealthError?
-        giniHealth.fetchDataForReview(documentId: MockSessionManager.missingDocumentID) { result in
-            switch result {
-            case .success(_):
-                receivedError = nil
-            case .failure(let error):
-                receivedError = error
-            }
-            expectation.fulfill()
-        }
-        waitForExpectations(timeout: 1, handler: nil)
-
-        // Then
-        XCTAssertNotNil(receivedError)
-    }
-
-    func testGetAllExtractions_Success() {
-        // Given
-        let fileName = "test_doctorsname"
-        let expectedExtractionContainer: GiniHealthSDK.ExtractionsContainer? = GiniHealthSDKTests.load(fromFile: fileName)
-        guard let expectedExtractionContainer else {
-            XCTFail("Error loading file: `\(fileName).json`")
-            return
-        }
-        let expectedExtractions: [GiniHealthSDK.Extraction] = ExtractionResult(extractionsContainer: expectedExtractionContainer).extractions
-
-        // When
-        let expectation = self.expectation(description: "Getting all extractions")
-        var receivedExtractions: [GiniHealthSDK.Extraction]?
-        giniHealth.getAllExtractions(docId: MockSessionManager.doctorsNameDocumentID) { result in
-            switch result {
-            case .success(let extractions):
-                receivedExtractions = extractions
-            case .failure(_):
-                receivedExtractions = nil
-            }
-            expectation.fulfill()
-        }
-        waitForExpectations(timeout: 1, handler: nil)
-
-        // Then
-        XCTAssertNotNil(receivedExtractions)
-        XCTAssertEqual(receivedExtractions?.count, expectedExtractions.count)
-    }
-
-    func testGetAlllExtractions_Failure() {
-        // When
-        let expectation = self.expectation(description: "Extraction failure")
-        var receivedExtractions: [GiniHealthSDK.Extraction]?
-        giniHealth.getAllExtractions(docId: MockSessionManager.failurePayableDocumentID) { result in
-            switch result {
-            case .success(let extractions):
-                receivedExtractions = extractions
-            case .failure(_):
-                receivedExtractions = nil
-            }
-            expectation.fulfill()
-        }
-        waitForExpectations(timeout: 1, handler: nil)
-
-        // Then
-        XCTAssertNil(receivedExtractions)
-    }
-
-    func testGetDoctorsNameExtractions_Success() {
-        // Given
-        let expectedDoctorName = "DR. SOMMER TEAM"
-
-        // When
-        let expectation = self.expectation(description: "Getting doctor name extractions")
-        var receivedDoctorExtraction: GiniHealthSDK.Extraction?
-        giniHealth.getAllExtractions(docId: MockSessionManager.doctorsNameDocumentID) { result in
-            switch result {
-            case .success(let extractions):
-                receivedDoctorExtraction = extractions.first(where: { $0.name == ExtractionType.doctorName.rawValue })
-            case .failure(_):
-                receivedDoctorExtraction = nil
-            }
-            expectation.fulfill()
-        }
-        waitForExpectations(timeout: 1, handler: nil)
-
-        // Then
-        XCTAssertNotNil(receivedDoctorExtraction)
-        XCTAssertEqual(receivedDoctorExtraction?.value, expectedDoctorName)
     }
 
     func testLoadClientConfigurationFromFile() {
@@ -559,87 +148,5 @@ final class GiniHealthTests: XCTestCase {
         XCTAssertNotNil(clientConfiguration)
         XCTAssertEqual(clientConfiguration.communicationTone, expectedDefaultComunicationTone)
         XCTAssertEqual(clientConfiguration.ingredientBrandType, expectedDefaultBrandType)
-    }
-
-    // MARK: - Delete Batch Of Documents Tests
-    /// Enum for Delete  Batch Document Types
-    private enum DeleteBatchDocumentType: String {
-        case unauthorizedDocuments
-        case notFoundDocuments
-        case missingCompositeDocuments
-        case success = ""
-        case failure
-
-        /// Expected failure response if applicable
-        var expectedFailure: [String]? {
-            switch self {
-            case .unauthorizedDocuments, .notFoundDocuments, .missingCompositeDocuments:
-                return [self.rawValue]
-            default:
-                return nil
-            }
-        }
-
-        /// Expected success response if applicable
-        var expectedSuccess: String? {
-            self == .success ? "" : nil
-        }
-    }
-
-    func testDeleteBatchDocumentsSuccess() {
-        performDeleteBatchDocumentsTest(
-            documentType: .success,
-            description: "Deleting Batch Of Documents with Success"
-        )
-    }
-
-    func testDeleteBatchDocumentsFailure() {
-        performDeleteBatchDocumentsTest(
-            documentType: .failure,
-            description: "Deleting Batch Of Documents with Failure"
-        )
-    }
-
-    func testDeleteBatchDocumentsErrorUnauthorizedDocuments() {
-        performDeleteBatchDocumentsTest(
-            documentType: .unauthorizedDocuments,
-            description: "Deleting Batch Of Documents with Error of unauthorized documents"
-        )
-    }
-
-    func testDeleteBatchDocumentsErrorNotFoundDocuments() {
-        performDeleteBatchDocumentsTest(
-            documentType: .notFoundDocuments,
-            description: "Deleting Batch Of Documents with Error of not found documents"
-        )
-    }
-
-    func testDeleteBatchDocumentsErrorMissingCompositeDocuments() {
-        performDeleteBatchDocumentsTest(
-            documentType: .missingCompositeDocuments,
-            description: "Deleting Batch Of Documents with Error of missing composite documents"
-        )
-    }
-
-    /// Helper Function for Delete Batch Documents Tests
-    private func performDeleteBatchDocumentsTest(
-        documentType: DeleteBatchDocumentType,
-        description: String
-    ) {
-        let expectation = self.expectation(description: description)
-        
-        let documentIds = documentType == .failure ? [] : [documentType.rawValue]
-        giniHealth.deleteDocuments(documentIds: documentIds) { result in
-            switch result {
-            case .success(let responseMessage):
-                XCTAssertEqual(responseMessage, documentType.expectedSuccess)
-            case .failure(let error):
-                let receivedError = error.unauthorizedDocuments ?? error.notFoundDocuments ?? error.missingCompositeDocuments
-                XCTAssertEqual(receivedError, documentType.expectedFailure)
-            }
-            expectation.fulfill()
-        }
-
-        waitForExpectations(timeout: 1, handler: nil)
     }
 }
